Add Error Checking and Additional Kernel Tests for new tensor shapes

+++ Cuda_kernels/Utils.cu
@@ -4,7 +4,7 @@
 
 using namespace std;
 
-Tensor random_fill(std::initializer_list<int> lst, float lo, float hi) {
+Tensor random_fill(std::initializer_list<unsigned> lst, float lo, float hi) {
 
   random_device               rd;
   mt19937                     gen(rd());
@@ -19,19 +19,19 @@ Tensor random_fill(std::initializer_list<int> lst, float lo, float hi) {
 
 Tensor
 cp4recom(Tensor FilterK, Tensor FilterC, Tensor FilterR, Tensor FilterS) {
-  const size_t rank = FilterK.shape[1];
-  const int    FK   = FilterK.shape[0];
-  const int    FC   = FilterC.shape[0];
-  const int    FR   = FilterR.shape[0];
-  const int    FS   = FilterS.shape[0];
-  Tensor       Out  = { FK, FC, FR, FS };
+  const unsigned rank = FilterK.shape[1];
+  const unsigned FK   = FilterK.shape[0];
+  const unsigned FC   = FilterC.shape[0];
+  const unsigned FR   = FilterR.shape[0];
+  const unsigned FS   = FilterS.shape[0];
+  Tensor         Out  = { FK, FC, FR, FS };
 
   // clang-format off
-  for (int a = 0; a < FK; ++a)
-  for (int b = 0; b < FC; ++b)
-  for (int c = 0; c < FR; ++c)
-  for (int d = 0; d < FS; ++d)
-  for (int r = 0; r < rank; ++r)
+  for (unsigned a = 0; a < FK; ++a)
+  for (unsigned b = 0; b < FC; ++b)
+  for (unsigned c = 0; c < FR; ++c)
+  for (unsigned d = 0; d < FS; ++d)
+  for (unsigned r = 0; r < rank; ++r)
     Out.m_data[a*FC*FR*FS + b*FR*FS + c*FS + d]
       += FilterK.m_data[a*rank + r]
        * FilterC.m_data[b*rank + r]
@@ -43,22 +43,22 @@ cp4recom(Tensor FilterK, Tensor FilterC, Tensor FilterR, Tensor FilterS) {
 }
 
 
-Tensor padNCHW(Tensor In, int pad = 1) {
-  int N  = In.shape[0];
-  int C  = In.shape[1];
-  int iH = In.shape[2];
-  int iW = In.shape[3];
+Tensor padNCHW(Tensor In, unsigned pad = 1) {
+  unsigned N  = In.shape[0];
+  unsigned C  = In.shape[1];
+  unsigned iH = In.shape[2];
+  unsigned iW = In.shape[3];
 
-  int oH = iH + (2 * pad);
-  int oW = iW + (2 * pad);
+  unsigned oH = iH + (2 * pad);
+  unsigned oW = iW + (2 * pad);
 
   Tensor Out = { N, C, oH, oW };
 
   // clang-format off
-  for (int n=0; n<N; ++n)
-  for (int c=0; c<C; ++c)
-  for (int h=0; h<oH; ++h)
-  for (int w=0; w<oW; ++w)
+  for (unsigned n=0; n<N; ++n)
+  for (unsigned c=0; c<C; ++c)
+  for (unsigned h=0; h<oH; ++h)
+  for (unsigned w=0; w<oW; ++w)
     if(h>=pad && h<=iH && w>=pad && w<=iW){
       Out.m_data[n*C*oH*oW + c*oH*oW + h*oW + w]
         = In.m_data[n*C*iH*iW + c*iH*iW + (h-pad)*iW + (w-pad)];
